pyserial_test/pyproject.toml
[tool.poetry]
name = "pyserial-test"
version = "0.1.0"
description = ""
authors = ["aagz <aagzip@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyserial = "^3.5"
pygame = "^2.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
